Use oks score for reprojections (#1836)
* Move compute_oks function to utils (avoid circular import)

* Update InstanceGroup score when update_points

* Add score to InstanceGroup repr

* Add score to InstanceGroup table

* (De)Serialize new score attribute

* Add more docs on score attribute

* Test score attribute
