+++ tests/test_util.py
@@ -154,3 +154,45 @@ def test_decode_preview_image(flies13_skeleton: Skeleton):
     img_b64 = skeleton.preview_image
     img = decode_preview_image(img_b64)
     assert img.mode == "RGBA"
+
+
+def test_compute_oks():
+    # Test compute_oks function with the cocoutils implementation
+    inst_gt = np.array([[0, 0], [1, 1], [2, 2]]).astype("float32")
+    inst_pr = np.array([[0, 0], [1, 1], [2, 2]]).astype("float32")
+    oks = compute_oks(inst_gt, inst_pr)
+    np.testing.assert_allclose(oks, 1)
+
+    inst_pr = np.array([[0, 0], [1, 1], [np.nan, np.nan]]).astype("float32")
+    oks = compute_oks(inst_gt, inst_pr)
+    np.testing.assert_allclose(oks, 2 / 3)
+
+    inst_gt = np.array([[0, 0], [1, 1], [np.nan, np.nan]]).astype("float32")
+    inst_pr = np.array([[0, 0], [1, 1], [2, 2]]).astype("float32")
+    oks = compute_oks(inst_gt, inst_pr)
+    np.testing.assert_allclose(oks, 1)
+
+    inst_gt = np.array([[0, 0], [1, 1], [np.nan, np.nan]]).astype("float32")
+    inst_pr = np.array([[0, 0], [1, 1], [np.nan, np.nan]]).astype("float32")
+    oks = compute_oks(inst_gt, inst_pr)
+    np.testing.assert_allclose(oks, 1)
+
+    # Test compute_oks function with the implementation from the paper
+    inst_gt = np.array([[0, 0], [1, 1], [2, 2]]).astype("float32")
+    inst_pr = np.array([[0, 0], [1, 1], [2, 2]]).astype("float32")
+    oks = compute_oks(inst_gt, inst_pr, False)
+    np.testing.assert_allclose(oks, 1)
+
+    inst_pr = np.array([[0, 0], [1, 1], [np.nan, np.nan]]).astype("float32")
+    oks = compute_oks(inst_gt, inst_pr, False)
+    np.testing.assert_allclose(oks, 2 / 3)
+
+    inst_gt = np.array([[0, 0], [1, 1], [np.nan, np.nan]]).astype("float32")
+    inst_pr = np.array([[0, 0], [1, 1], [2, 2]]).astype("float32")
+    oks = compute_oks(inst_gt, inst_pr, False)
+    np.testing.assert_allclose(oks, 1)
+
+    inst_gt = np.array([[0, 0], [1, 1], [np.nan, np.nan]]).astype("float32")
+    inst_pr = np.array([[0, 0], [1, 1], [np.nan, np.nan]]).astype("float32")
+    oks = compute_oks(inst_gt, inst_pr, False)
+    np.testing.assert_allclose(oks, 1)
